Remove requirement for separately build container image
As it stands this pipeline runs against the old version of the
container while the image is being rebuilt meaning that changes to
the dependencies are not applied when the tests are run.

This recently broke because we merged a dependabot advisory which
got through the pipeline untested and broke production.